Javascripty e ajustes de components
