--- Public/components/table-row.php
@@ -1,15 +1,0 @@
-<tr>
-
-    <td>ABC-1234</td>
-    <td>Santana CG</td>
-    <td class='d-none d-sm-table-cell'>VW</td>
-    <td class='d-none d-sm-table-cell'>1986</td>
-    <td>
-        <a href="#editModal" class="edit" data-bs-toggle="modal">
-            <i class="material-icons" data-toggle="tooltip" title="Edit">&#xE254;</i>
-        </a>
-        <a href="#deleteModal" class="delete" data-bs-toggle="modal">
-            <i class="material-icons" data-toggle="tooltip" title="Delete">&#xE872;</i>
-        </a>
-    </td>
-</tr>
